Update vav4inject.js
removed fly & InfFly as it is patched due to Vector's custom anticheat. It was only useful on some servers such as (oneblock). ._.

The only way its possible to make it work again is making a desync 2BPS fly. 

How this desync works:

- By not incrementing inputSequenceNumber while desynced
keep the position diff magnitude < 2 (clamp to like 1.98 or 1.99, whichever works best without flagging)

- And to know what your position is on the server, listen to the reconciliation (I think that's what it was called lol) packet and the position stored there is your position on the server.

- this is from MTL (Miniblox Transition Layer)