Update README.md
i think we should remove fastbreak as it is broken and not used for anything? it saves lines of code so we can focus on important modules. And removed .toggle all from know issues. i will not be using Bugs_And_Fixes.md as i will move it all to this table in the readme markdown (in known issues!) 

Signed-off-by: TheMiddleMan <192749026+TheM1ddleM1n@users.noreply.github.com>